requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-generativeai==0.8.3
pdfplumber==0.10.3
pypdf2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6